Use hypervisor interface to render virt-launcher pod spec
