+++ pkg/hypervisor/kvm/BUILD.bazel
@@ -1,0 +1,18 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = ["launcherresourcerenderer.go"],
+    importpath = "kubevirt.io/kubevirt/pkg/hypervisor/kvm",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//pkg/downwardmetrics:go_default_library",
+        "//pkg/tpm:go_default_library",
+        "//pkg/util:go_default_library",
+        "//pkg/util/hardware:go_default_library",
+        "//staging/src/kubevirt.io/api/core/v1:go_default_library",
+        "//staging/src/kubevirt.io/client-go/log:go_default_library",
+        "//vendor/k8s.io/api/core/v1:go_default_library",
+        "//vendor/k8s.io/apimachinery/pkg/api/resource:go_default_library",
+    ],
+)
