+++ pkg/virt-controller/services/BUILD.bazel
@@ -18,10 +18,10 @@ go_library(
         "//pkg/apimachinery:go_default_library",
         "//pkg/config:go_default_library",
         "//pkg/container-disk:go_default_library",
-        "//pkg/downwardmetrics:go_default_library",
         "//pkg/dra:go_default_library",
         "//pkg/hooks:go_default_library",
         "//pkg/host-disk:go_default_library",
+        "//pkg/hypervisor:go_default_library",
         "//pkg/network/downwardapi:go_default_library",
         "//pkg/network/istio:go_default_library",
         "//pkg/network/multus:go_default_library",
