+++ pkg/virt-controller/services/template.go
@@ -46,6 +46,7 @@ import (
 	"kubevirt.io/client-go/precond"
 
 	drautil "kubevirt.io/kubevirt/pkg/dra"
+	"kubevirt.io/kubevirt/pkg/hypervisor"
 	"kubevirt.io/kubevirt/pkg/pointer"
 
 	"kubevirt.io/kubevirt/pkg/apimachinery"
@@ -77,13 +78,12 @@ const (
 	virtExporter     = "virt-exporter"
 )
 
-const KvmDevice = "devices.kubevirt.io/kvm"
-const TunDevice = "devices.kubevirt.io/tun"
-const VhostNetDevice = "devices.kubevirt.io/vhost-net"
-const SevDevice = "devices.kubevirt.io/sev"
-const VhostVsockDevice = "devices.kubevirt.io/vhost-vsock"
-const PrDevice = "devices.kubevirt.io/pr-helper"
-
+const K8sDevicePrefix = "devices.kubevirt.io"
+const TunDevice = K8sDevicePrefix + "/tun"
+const VhostNetDevice = K8sDevicePrefix + "/vhost-net"
+const SevDevice = K8sDevicePrefix + "/sev"
+const VhostVsockDevice = K8sDevicePrefix + "/vhost-vsock"
+const PrDevice = K8sDevicePrefix + "/pr-helper"
 const debugLogs = "debugLogs"
 const logVerbosity = "logVerbosity"
 const virtiofsDebugLogs = "virtiofsdDebugLogs"
@@ -114,11 +114,6 @@ const EXT_LOG_VERBOSITY_THRESHOLD = 5
 const ephemeralStorageOverheadSize = "50M"
 
 const (
-	VirtLauncherMonitorOverhead = "25Mi"  // The `ps` RSS for virt-launcher-monitor
-	VirtLauncherOverhead        = "100Mi" // The `ps` RSS for the virt-launcher process
-	VirtlogdOverhead            = "25Mi"  // The `ps` RSS for virtlogd
-	VirtqemudOverhead           = "40Mi"  // The `ps` RSS for virtqemud
-	QemuOverhead                = "30Mi"  // The `ps` RSS for qemu, minus the RAM of its (stressed) guest, minus the virtual page table
 	// Default: limits.memory = 2*requests.memory
 	DefaultMemoryLimitOverheadRatio = float64(2.0)
 
@@ -368,7 +363,7 @@ func (t *TemplateService) renderLaunchManifest(vmi *v1.VirtualMachineInstance, i
 	if err != nil {
 		return nil, err
 	}
-	resourceRenderer, err := t.newResourceRenderer(vmi, networkToResourceMap)
+	resourceRenderer, err := t.newResourceRenderer(vmi, networkToResourceMap, t.clusterConfig.GetHypervisor().Name)
 	if err != nil {
 		return nil, err
 	}
@@ -405,6 +400,7 @@ func (t *TemplateService) renderLaunchManifest(vmi *v1.VirtualMachineInstance, i
 			"--hook-sidecars", strconv.Itoa(len(requestedHookSidecarList)),
 			"--ovmf-path", ovmfPath,
 			"--disk-memory-limit", strconv.Itoa(int(t.clusterConfig.GetDiskVerification().MemoryLimit.Value())),
+			"--hypervisor", t.clusterConfig.GetHypervisor().Name, // TODO This might need to be removed because virt-launcher is no longer updated
 		}
 		if nonRoot {
 			command = append(command, "--run-as-nonroot")
@@ -901,11 +897,12 @@ func (t *TemplateService) newVolumeRenderer(vmi *v1.VirtualMachineInstance, imag
 	return volumeRenderer, nil
 }
 
-func (t *TemplateService) newResourceRenderer(vmi *v1.VirtualMachineInstance, networkToResourceMap map[string]string) (*ResourceRenderer, error) {
+func (t *TemplateService) newResourceRenderer(vmi *v1.VirtualMachineInstance, networkToResourceMap map[string]string, hypervisorName string) (*ResourceRenderer, error) {
 	vmiResources := vmi.Spec.Domain.Resources
+	hypervisorDeviceK8sResource := K8sDevicePrefix + "/" + hypervisor.NewLauncherResourceRenderer(hypervisorName).GetHypervisorDevice()
 	baseOptions := []ResourceRendererOption{
 		WithEphemeralStorageRequest(),
-		WithVirtualizationResources(getRequiredResources(vmi, t.clusterConfig.AllowEmulation())),
+		WithVirtualizationResources(getRequiredResources(vmi, hypervisorDeviceK8sResource, t.clusterConfig.AllowEmulation())),
 	}
 
 	if err := validatePermittedHostDevices(&vmi.Spec, t.clusterConfig); err != nil {
@@ -1260,29 +1257,6 @@ func appendUniqueImagePullSecret(secrets []k8sv1.LocalObjectReference, newsecret
 	return append(secrets, newsecret)
 }
 
-func addProbeOverheads(vmi *v1.VirtualMachineInstance, quantity *resource.Quantity) {
-	// We need to add this overhead due to potential issues when using exec probes.
-	// In certain situations depending on things like node size and kernel versions
-	// the exec probe can cause a significant memory overhead that results in the pod getting OOM killed.
-	// To prevent this, we add this overhead until we have a better way of doing exec probes.
-	// The virtProbeTotalAdditionalOverhead is added for the virt-probe binary we use for probing and
-	// only added once, while the virtProbeOverhead is the general memory consumption of virt-probe
-	// that we add per added probe.
-	virtProbeTotalAdditionalOverhead := resource.MustParse("100Mi")
-	virtProbeOverhead := resource.MustParse("10Mi")
-	hasLiveness := vmi.Spec.LivenessProbe != nil && vmi.Spec.LivenessProbe.Exec != nil
-	hasReadiness := vmi.Spec.ReadinessProbe != nil && vmi.Spec.ReadinessProbe.Exec != nil
-	if hasLiveness {
-		quantity.Add(virtProbeOverhead)
-	}
-	if hasReadiness {
-		quantity.Add(virtProbeOverhead)
-	}
-	if hasLiveness || hasReadiness {
-		quantity.Add(virtProbeTotalAdditionalOverhead)
-	}
-}
-
 func HaveContainerDiskVolume(volumes []v1.Volume) bool {
 	for _, volume := range volumes {
 		if volume.ContainerDisk != nil {
@@ -1571,7 +1545,8 @@ func CalculateMemoryOverhead(clusterConfig *virtconfig.ClusterConfig, netBinding
 	if vmiCPUArch == "" {
 		vmiCPUArch = clusterConfig.GetClusterCPUArch()
 	}
-	memoryOverhead := GetMemoryOverhead(vmi, vmiCPUArch, clusterConfig.GetConfig().AdditionalGuestMemoryOverheadRatio)
+
+	memoryOverhead := hypervisor.NewLauncherResourceRenderer(clusterConfig.GetHypervisor().Name).GetMemoryOverhead(vmi, vmiCPUArch, clusterConfig.GetConfig().AdditionalGuestMemoryOverheadRatio)
 
 	if netBindingPluginMemoryCalculator != nil {
 		memoryOverhead.Add(
