+++ tests/migration/BUILD.bazel
@@ -23,6 +23,7 @@ go_library(
         "//pkg/certificates/triple:go_default_library",
         "//pkg/certificates/triple/cert:go_default_library",
         "//pkg/controller:go_default_library",
+        "//pkg/hypervisor:go_default_library",
         "//pkg/libdv:go_default_library",
         "//pkg/libvmi:go_default_library",
         "//pkg/libvmi/cloudinit:go_default_library",
@@ -31,7 +32,6 @@ go_library(
         "//pkg/util:go_default_library",
         "//pkg/util/hardware:go_default_library",
         "//pkg/virt-config/featuregate:go_default_library",
-        "//pkg/virt-controller/services:go_default_library",
         "//pkg/virt-controller/watch/topology:go_default_library",
         "//pkg/virt-handler:go_default_library",
         "//pkg/virt-handler/cgroup:go_default_library",
